+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3pipelines 0.4.0-9000
 
 * `PipeOpLearner` `packages` slot is set to the `Learner`'s `packages`
+* Bugfix: `PipeOp` `train()` and `predict()` report correct channel name when output has wrong type.
 
 # mlr3pipelines 0.4.0